refactor(scan): split tokeignore template owner
