+++ docs/architecture-consolidation-plan.md
@@ -86,6 +86,7 @@ fixtures:
 | Sensor command | `crates/tokmd/src/commands/sensor.rs` and `crates/tokmd/src/commands/sensor/` | `sensor.rs` 157; findings owner/tests 344; gate mapping owner/tests 189; output owner/tests 141 | Sensor finding emission, cockpit-evidence-to-envelope gate mapping, and 3-layer output topology now live in owner modules while the command keeps git resolution, cockpit execution, and envelope assembly. |
 | Model aggregation | `crates/tokmd-model/src/lib.rs`, `crates/tokmd-model/src/aggregate.rs`, `crates/tokmd-model/src/children.rs`, `crates/tokmd-model/src/rows.rs`, and `crates/tokmd-model/src/sorting.rs` | `lib.rs` 255; aggregation owner/tests 344; child-language owner/tests 160; row collection owner/tests 411; sorting owner/tests 91 | Report builders, file-row collection, child/embedded language aggregation, in-memory row detection, and row sorting now live in owner modules; future model work should start from concrete product or proof evidence rather than splitting for line count |
 | Scan root handling | `crates/tokmd-scan/src/lib.rs`, `crates/tokmd-scan/src/roots.rs`, and `crates/tokmd-scan/src/path/` | scan facade 534; root validation/rebasing owner 62; bounded path primitives/tests under `path/` | Keep public scan/config/in-memory facades and ignore-pattern expansion in `lib.rs`; keep validated root construction and caller-facing report-path rebasing in `roots.rs`; keep bounded relative/root path primitives under `path/`. Preserve the `model_scan_path_normalization` proof scope. |
+| Scan init templates | `crates/tokmd-scan/src/tokeignore/mod.rs` and `crates/tokmd-scan/src/tokeignore/templates.rs` | init coordinator 131; template catalog owner/tests 261 | Keep CLI-independent `.tokeignore` file creation, force behavior, and print behavior in the coordinator; keep profile template selection and template invariants in `templates.rs`. Preserve the `tokmd_init_tokeignore` proof scope. |
 | Scan walk Git boundary | `crates/tokmd-scan/src/walk/mod.rs` and `crates/tokmd-scan/src/walk/git.rs` | walk facade 341; git listing/bounding owner 141 | Keep filesystem and MemFs traversal in `walk/mod.rs`; keep git-backed listing, subprocess environment scrubbing, and tracked-file path bounding in `walk/git.rs`. Preserve the `git_subprocess_boundary` proof scope. |
 
 ## Batch Order
